backend/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
pyjwt>=2.8.0
reportlab>=4.0.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
